+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/General/Axes/XAxisTicksFormatted.jsx
@@ -1,0 +1,37 @@
+import React from 'react';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
+import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
+import '@cerner/terra-graphs/lib/components/Graph.module.scss';
+import lineTimeseries from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineTimeseries';
+
+const tickValues = [
+  new Date(2016, 0, 1, 1, 0).toISOString(),
+  new Date(2016, 0, 1, 5, 0).toISOString(),
+  new Date(2016, 0, 1, 10, 0).toISOString(),
+  new Date(2016, 0, 1, 15, 0).toISOString(),
+  new Date(2016, 0, 1, 20, 0).toISOString(),
+];
+
+/*
+Please refer to the documentation below to see the graphConfig and data objects
+*/
+const graphConfig = utils.deepClone(lineTimeseries('#xAxisTicksFormatted'));
+graphConfig.axis.x.ticks = {
+  values: tickValues,
+  format: '%a %b %e %X %Y',
+};
+
+const GeneralExample = () => {
+  React.useEffect(() => {
+    Carbon.api.graph(graphConfig);
+  }, []);
+  return (
+    <React.Fragment>
+      <div id="tooltip" className="initial-tooltip" />
+      <div id="xAxisTicksFormatted" />
+    </React.Fragment>
+  );
+};
+
+export default GeneralExample;