+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/General/Padding/CustomLegendPlacement.jsx
@@ -1,0 +1,37 @@
+import React from 'react';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
+import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
+import '@cerner/terra-graphs/lib/components/Graph.module.scss';
+import '@cerner/terra-graphs/lib/components/Line/LineGraph.module.scss';
+import lineTimesries from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineTimeseries';
+import exampleData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/multiTimeseriesData';
+
+/*
+Please refer to the documentation below to see the graphConfig and data objects
+*/
+const graphConfig = utils.deepClone(lineTimesries('#graphContainer'));
+graphConfig.bindLegendTo = '#legendContainer';
+const dataset = utils.deepClone(exampleData);
+
+const GeneralExample = () => {
+  React.useEffect(() => {
+    const graph = Carbon.api.graph(graphConfig);
+    dataset.forEach((data) => {
+      graph.loadContent(Carbon.api.line(data));
+    });
+  }, []);
+  return (
+    <div id="customLegendPlacement">
+      <div id="tooltip" className="initial-tooltip" />
+      <div className="bindto-container">
+        <div id="legendContainer" className="legend-bindto-container" />
+        <div id="graphContainer" className="legend-bindto-graph-container" />
+      </div>
+      <div id="graphContainer" />
+    </div>
+  );
+};
+
+export default GeneralExample;
+